Separate models into module.
