+++ real_intent/events/__init__.py
@@ -1,4 +1,4 @@
 """Event generation."""
-from real_intent.events.base import EventsResponse
+from real_intent.events.models import EventsResponse
 from real_intent.events.perplexity import PerplexityEventsGenerator
 from real_intent.events.scrapy_claude import ScrapybaraEventsGenerator
