+++ real_intent/events/scrapy_claude/__init__.py
@@ -19,7 +19,8 @@ from real_intent.events.scrapy_claude.claude_sync import (
     ToolCollection, 
     ComputerTool
 )
-from real_intent.events.base import Event, EventsResponse, BaseEventsGenerator
+from real_intent.events.models import Event, EventsResponse
+from real_intent.events.base import BaseEventsGenerator
 from real_intent.events.errors import NoValidJSONError, NoEventsFoundError
 from real_intent.events.utils import extract_json_only, retry_generation
 from real_intent.internal_logging import log
